requirements.txt
dspy-ai
python-dotenv
pymongo[srv]
firebase-admin
newspaper3k
requests
anthropic
fastapi
uvicorn
python-socketio
python-multipart
pymupdf
pymupdf4llm
Pillow
ragatouille
lxml-html-clean
paramiko
motor